colourpartition.c: the invariant is more complex than originally thought.
The sequence of known-to-be-white at the start has to be ended by a
known-to-be-non-white value.  Likewise, the sequence of known-to-be-red
at the end has to be ended by a known-to-be-non-red value.  The initial
loops that set the invariants have to be used again after each swap.
This is a working version; it may be possible to put the invariant loops
as first steps in the main loop - next experiment.