Add answer for SO 20889460 and move binary.c for SO 20342745.
SO 20889460 is an older question, but given a new answer that is
modestly clever, and earned a revival badge (answer after 30 days and
first answer to get two upvotes).  One existing answer gained a
down-vote after my answer was added, and it wasn't my down-vote.

SO 20342745 is also an old question, but my code's been lurking around
for an aeon or so.