Upgrade 3dalloc[34].c with argument handling
Note that the answer has multiple versions of the code.  It seems that
3dalloc1.c was known to be leaky. 3dalloc2.c was a first flawed attempt
to fix the leal. 3dalloc3.c did the job properly. 3dalloc4.c
dramatically reduces the number of allocations.  However, because there
was no annotation in the code, this was not self-evident 2.75 years
later.  So, temporarily, 3dalloc2.c and 3dalloc3.c are identical.  The
original flaw will be reinstated and annotated later.  It pays to read
the answer (SO 1857-9583) before tampering.