src/matrices: increase size of matrices in matmul89.c, matmul99.c
Change from multiplying a 2x2 and a 2x1 matrix (yielding a 1x2 matrix)
to multiplying a 3x4 and a 4x2 matrix (yielding a 3x2 matrix).
Adjust formatting slightly.  The matrices are all allocated as
regular (non-VLA) arrays - but the manipulation in matmul99.c
uses VLAs in the functions.  The programs use the same data
and produce the same result, but the processing is somewhat
different, of necessity.