SO 4178-6918 (Quick Sort Again)
Using Wikipedia, it was straight-forward to code the algorithm.  Need to analyze the pseudo-code in the question for accuracy.  One known bug is that the code in the question does not iterate where it should. Another is that it chooses the end rather than the start as the pivot - well, that might be questionable.  Would also like to lose the infinite loop.