LIBSOQ: stderr.c - fewer split messages
Integrate changes from main code line for RCS version 10.18.  Uses
err_fmtmsg() to create a single string containing the message to be
printed (in the main code, even for syslog).  Then does a single write.
The code was already around, more or less, for file descriptor and
syslog handling; now it is used for standard I/O too.  The multi-child
program in src/so-4319-0158 seems to demonstrate that the fix does its
job, at least most of the time.