SO 4404-0027: Add -t trigger to trace trigger words.
It appears that the trie incorrectly reports words are present.  In
particular, it was noted that when the dictionary contained
'journeyings' and not 'journeying' but the text contained 'journeying',
then this was recognized as a word present in the dictionary.  This is
not good.  It isn't clear to me any more (actually, it never really was
clear to me) how the dictionary knows whether a word ends at a
particular point.  There needs to be a way to print all the words in the
trie (dictionary) so that the list can be validated against the
dictionary.  The fix probably requires a complete rewrite of the trie
creation (and hence the search code too).