SO 4404-0027: Remove -t option and debug printing.
Dump the trie when the debug option is used, but only dump words, not
prefixes.  Seems to be working in synchrony with the OP's code which
does dynamic allocation of the data trie data.  This runs faster than
that — surprise, surprise.  But not by as much as might be expected.
I think that the wrong word handling is clumsy, creating a new entry
with word and offset each time a wrong word is found.  The code should
identify how much of the trie is in use — there's probably a lot not
used.  Similarly with the wrong word list, though that probably needs
restructuring as a second trie, with a list of offsets for each word.
That is mostly a problem for the OP.