SO 4404-0027: Add spell-checker-ac.c
According to OP, the online judge accepted output from spell-checker-ac.
Since it generates the same output as sc59.c on the same inputs, this is
good news.  The claim is it is slow.  That's possible.  Measurements
show that sc59.c spends about 0.139 seconds building the trie, about
0.226 seconds scanning the main document, and about 0.005 seconds
printing the results.  That's interesting; I suspected the printing
would take longer.  Clearly, a major preoccupation should be speeding up
the scan process.  The trie size (1,500,000) is not too grotesquely big;
about 1,126,000 entries are used.