SO 4479-2794: Fix compilations on Ubuntu 16.04 LTS
Mostly; the default make setting of ARFLAGS is broken when ar is compiled with --enable-deterministic-archives, which is a pusillanimous choice of option. If people want to break things, make so you have to actively break it, rather than requiring people to stop breaking it.  Hate, loathe, hate, loathe!  In case you missed it -- I think this is an atrocious choice and I would not do it.  Also handle getcwd() return value cannot be ignored.