rfn.sh - report name once when appropriate
If the standard output is a terminal, only report the name once.  If the standard output is not a terminal, assume it is being used in a context such as "vim $(rfn-c base)", sending one copy to the command substitution on standard output and one to standard error to inform the user.