SO 4545-9691: Finding prime pairs
Material not submitted to SO.  Based on observations that all prime pairs after {3, 5} satisfy 6N±1 for some integral N.
Also using further observation that for N of the form 5M±1, one of the two numbers in the prime pair is a multiple of 5.
This trims the number of pairs reviewed to 60% of the pairs produces by 6N±1.  The factorization only needs
to start at 7 — by construction, none of 2, 3, or 5 is a factor.