libcs50: upgrade to newer code
Add diatribe to README.md about the modified interface,
which is grotesquely out of order - it is not code that the students could be expected to write,
or should be expected to understand.  It has preprocessor wizardry that is very complex,
and the side effects are pretty weird.  Also, you can no longer free the strings returned
by get_string() (or GetString()) because there is a destructor installed to do that automatically.
All in all, the upgrade is not appropriate.