Add LICENSE.md
Summary points out that all material on SO is
licenced under CC BY-SA 3.0 - with attribution - so this code is likewise, but
offers the GPL v3 as an alternative.  Since the GPL v3 is more stringent, it is
not a particularly sensible option to choose, but that is an SEP - Someone Else's Problem.
There is a Markdown version of the CC BY-SA 3.0 text included, but the online
version is better formatted - Markdown does not really encourage alphabetic
bullet points or lower-case Roman numerals, both of which are used by the
online version.