Add support for Perl scripts
Even now, (GNU) Make does not have support for Perl scripts by default.
Add rules to etc/soq-head.mk to handle Perl.  The rule uses the fixin
script, the source for which is in src/scripts/fixin.pl.  It is assumed
that fixin is on your PATH.  There's a chicken and egg bootstrapping
problem - use:

    make FIXIN='perl fixin.pl'

to get over the bootstrap hump.  This is not explicitly documented in
the src/scripts/makefile.  Of course, the local variant of Make, RMK,
supports Perl scripts out of the box, so it isn't a problem, which is
why it was noticed until using GNU Make instead of RMK on a Windows
machine (under Cygwin).