SO 4578-8729: bare minimum changes to compile
Roughly the bare minimum changes to get the code to compile under my default compilation options.  Three functions are declared (extern), rather than made static; that might be argued to be "not minimal".  The change of type in the argument to getWord() is a question of self-consistency.  Fixing conversion specifications to match the type is simply a necessity.