SO 4305-9874: Document pedantic incorrectness.
The code in pd37.c and pd47.c does not compile with the -pedantic flag because the code passes pointers to functions to code expecting pointer to void, and pointers to object and pointers to functions are not strictly compatible with each other.  Sometime, I may go back and see whether the discrepancy can be fixed, perhaps
replacing void * with void (*)(void) in some places.