SO 4829-1562: Add timing test
NB: got a segfault on quick_sort() for 1000x1000.
Not the right time to investigate why - but it is puzzling.
However, it was clear that the copy-quicksort-copy code is
dramatically faster than the non-copying sort.  The benefit
is marginal on 10x10 (15µs vs 12µs), but 20x20 shows 172µs
vs 32µs in favour of quicksort, and by 900x900, the basic
sort was taking 285s vs 0.054s for quicksort.