src/scripts: Fettle pipe-rot.sh again.
For better or worse, SO comments are limited to 600 characters.
Using pipe-rot -c generates 595 characters (including a trailing space)
which can be used.  Using pipe-rot -a or with no arguments generates
a slightly less compressed version.  There is room to debate whether
the URL for dup() is needed; it is the same page as the dup2() URL.