SO 3443-6734: Fix major bugs in rand_int() and the shuffle.
The random integer code wasn't working as intended.  The shuffle code
completely flubbed the loop, decrementing n instead of i as intended.
Problem spotted when using a transcription of the code for another
problem.  Note that the Wikipedia article has a discussion on common
implementation errors, and notes that the random integer problem
can lead to problems where the shuffle never leaves an element in its
original position — the problem that triggered some reconsideration.