src/sortfile: update annotation on leak prevention
The leak occurs from the bowels of getdelim (getline), and isn't the
final buffer allocated by the call that gets EOF (that is explicitly
freed).

etc/suppressions-macos-10.13.4: add a stray suppression for the pre-main
startup code and also a suppression for the getdelim/getline issue.

Keep the annotations in the src/sortfile/sortfile1.c and
src/sortfile/sortfile2.c files so that they can be rechecked in a due
course.