Basic searching for sequences of words OK.
Per design requirements, it requires no spaces or punctuation in the
word sequence, so "onethousandtwohundredandtwentysix" works as a word
sequence, but add a space in it and it will fail.

Add alt.words.03 which contains some compound word sequences.

Add words.numeric which includes ordinals (zeroth, first, etc).

Update trie89.c to print a complaint if it fails to open fixed file
name words.list.  (Next change is to process names given on command
line, complaining if there are none.)