SO 3994-9116: Upgrades to trie-search13.c
Add some more compound number words to alt.words.03

Upgrade trie-search13.c to: (a) require a dictionary to be loaded before
searching for words and (b) accumulate number words in an array and
print them.  Item (b) is a precursor to 'find words and non-words in the
input', and then convert words to numbers.  This will use the
number-to-words and words-to-numbers code from LibJL, probably as
amended.

To be considered: 'three gross', 'four score and ten', 'two dozen' and
sundry others.  Should the Indian lakh, crore, etc system be integrated?
Should there be a way to control the presence (English) or absence
(American) of 'and' in the spelling of a number?  Should 'a' and 'an'
(and 'o' and 'oh') be allowed as number words?  (Currently, 'a' is but
none of 'an', 'o' and 'oh' is allowed.)

What is the goal of the system?  To be able to take undivided verbiage
such as 'adozenpeopleandtheirthreescoredonkeysateonegrossofbagels' into
'a dozen people and their three score donkeys ate one gross of bagels'.
However, it would be acceptable to attain 'a dozen people and their
three score donkeysate one gross ofbagels', or even 'a dozen
peopleandtheir three score donkeysate one gross ofbagels'.  Also
'inthethirdyearofthereignofcharlesiofspaininthefifteenhundredandnineteenthyearofourlord'
might become 'in the third year of the reign of charles i of spain, in
the fifteen hundred and nineteenth year of our lord'.