SO 1886-3184: Add alternative test `badprobe.c`
The new probe tries to use `/dev/null` instead of a pipe.
It turns out that this does not work; the driver for `/dev/null`
does not do parameter checking that is crucial to the operation
of `memprobe.c`.

Expand README.md to discuss this and show a 'real world' piece
of code where `memprobe_rw()` might be used to good benefit.