SO 4997-3644: Create trie-split67.c
Copy trie-split67.c from trie-search13.c and modify it to use thhe code
from aoscopy.[ch] that's already in libsoq (unlike the aosptr.[ch] or
aomcopy.[ch] or aomptr.[ch] files).

In due course, it will recognize the 'unknown words' part of the input
string, and will split strings with blanks and punctuation (and digits)
into 'words' and only pass those to be analyzed to the word search code.