Revised to handle more or less arbitrary text.
The dictionary loading phase is fussy; it insists on pure alphabetic
data.

The word analysis phase is noisy but accepts interleaved numeric words
(assuming the dictionary is words.numeric) and non-words and non-alpha.
It processes README.md, for example, without baulking (though it is
very noisy about it).  The next phase will be some cleanup to reduce
the noise now that the code is more or less working.