SO 0143-3204: Move note on extern function declarations.
Gabriel Staples made a change in April 2018 which more or less copied a
comment from later in the answer to just before the first example
program.
This change had not previously been reflected in Git.
Making the change by moving the text (rather than repeating it), with
minor differences (the code does not contain the

    // "extern" is optional; see note above

comments that Gabriel added).
Also mention that C99 and C11 require functions to be declared or
defined before they are used.
It doesn't go into the intricacies of prototypes vs non-prototypes, etc;
that's a matter for a separate question.