SO 5100-4470: Why it is important to close unused pipe file descriptors.
Code showing why it is important to close unused ends of pipes, both the
read and the write ends of the pipe.

When invoked without any arguments, the programs (`fork13` and `fork29`)
fail to complete; one of the processes is jammed because it is either
reading or writing on a pipe where the only process with the
complementary end of the pipe open is the same process.

When invoked with arguments, the pipe descriptors are properly closed
and the pipeline completes (not always successfully — the writing
process gets a SIGPIPE signal — but without jamming indefinitely).