src/data-files: Upgrade us-name-collection.sh
Provide guidance on the home-brew nbncl, extags, unhtml commands.
They aren't currently in the SOQ repository, but there's no major
reason they couldn't be added.

The curl and column commands are not something for which there's
a simple surrogate — though column could probably be simulated
with awk if absolutely necessary.