Add some notes about what might be wanted.
It isn't 100% clear a priori what the results should be with mixed signs
on the operands.  If the quotient is -14.75, should rounding up go to
-15 or -14?  You can make a case for either.  Maybe -15 is 'round away
(from zero)' and the other is 'round up'.

Round nearest is also an option, and there is code for that (which may
not be working, but that's a separate issue).  It's mostly straight
forward except perhaps when the quotient is -14.50; should that round
towards +∞, to -∞, towards 0, to even, to odd.  Most, if not all, of
these have been used in some context or another.  Presumably, there
should be no difference between positive/negative and negative/positive?

Unsigned integers avoid most of the trickiness, of course!