SO 5256-3870: Refinements and additions.
In `array-range-19.c`, minor tweaks disclosed as necessary by what is
now `array-range-29.c`.  Program `array-range-29.c` can be compiled with
any of a large number of options to control the type.  Any of
`-DUSE_INTMAX`, `-DUSE_LONG_LONG`, `-DUSE_LONG`, `-DUSE_INT64`,
`-DUSE_INT32`, `-DUSE_INT16` or `-DUSE_INT`, (but the simply is
different from any of the others) gives a valid set of types and ranges.

In `array-range-23.c`, the parameterization is improved, consistent with
`array-range-19.c`.