src/Primes/isprime.c: Add extra prime tester isprime4().
Surprisingly, isprime2() now seems to be fastest, faster even than
isprime3(), and isprime4() is about the same as isprime3().  The
differences between isprime[1-4]() are fairly small — time range
currently 12.5-14.7, but the delta between 2 and the others is just
under 1 second, and the difference between the others is less than 0.1
second.  The actual times and number of primes found depends on the
seed.