Partially modified euler-08.c; added euler-08-2.c.
Trying to impove diagnostics in euler-08.c, but running into headwinds,
so rebuild the loop, with careful diagnostics from the start, in
euler-08-2.c.  The product_digits.sh script calculates the product of
digits in each line of input.  The output from euler-08-2.c looks good.
Now to capture maximum product and report at the end.