First pass at SO 5332-099 (find closest prime).
This code brute forces the search up and down, finding the closest prime
larger and smaller, and choosing whichever it is.

An alternative, more subtle algorithm would get onto the 6N±1 track upwards
and downwards of the starting point, and go around testing those values for
primality, working one step at a time to find the minimum distance prime(s).

What's here works, and it was easy to slot the binary search functions from
from bsmnmx.c into the code.