Marginally cleaned up.
As noted in email:

I'm not happy to see:

178             strcpy((states[currentStates]) -> code, foundCode);
179             states[currentStates] -> num_records=1;

There should never be spaces around an arrow '->' or dot '.' operator.
They bind too tight to warrant any space.

I'm not happy to see:

246             (*(states +rankOfState))->hi_millitime += *currentGeol;

It is so much easier to type and read states[rankOfState]->hi_millitime.
That's 6 punctuation characters instead of just 2, all else apart.  It
is also (much more importantly) easier to decipher.

I'm surprised to see the huge block of repeated code in the function;
that doubles the opportunities for bugs, and makes it harder to read,
again.

If you use a pointer in the form stateptr->hi_millitime or equivalent
(having set stateptr = states + rankOfState;), then pointers are easy
and useful.  If you're adding stuff to pointers before dereferencing
them, use the array-subscript notation.  Don't use the (*(states +
rankOfState))->hi_millitime notation.

In the first block, where you're dealing with a new state, you aren't
assigning many values to the state record.  You need to fix that.

In the second block, you're assigning the limit temperatures to

282                 states[currentStates]->lo_temp_reading = TEMP;

283                 states[currentStates]->lo_temp_timestamp = *currentTimeStamp;

but the rest of the code is assigning to:

271             (*(states +rankOfState))->pressure += PRESSURE;

That means you're pointing to the wrong place.  Consistency is so
important in programming.  When things are chaotic, there are usually
bugs.

You have some serious clean-up work to do.  You need to get rid about
half the code.  Decide which element of the states array needs to be
updated, and then use that the same regardless of whether the state was
seen before or not.