Remove redundant declarations; implement unimplemented.
Accessor functions list_head and list_tail were declared in the source
file (but not in the header), but neither implemented nor used.  Now
implemented.  The destructor and constructor were declared in the source
file (and implemented and used) as well as in the header; remove the
declarations form the source file.