Min-Reprex is an abomination!
Update scripts/sscce.sh to recognize the 'reprex' fiasco.

See the incorrectly titled [Min-Reprex: A Less Awkward Name For
MCVE](https://meta.stackoverflow.com/questions/384994/); the new name is
not an improvement at all (it is a ***MORE*** awkward name than MCVE,
and the change was mostly if not completely unnecessary.  The problem
with not finding MCVE via search on SO should have been fixed by using a
sensible search tool — the new name is confusing with the R package of
the same name, and generally sounds and looks horrid.  It is a black
mark against the SO team that this abomination was foisted onto the
world without consultation..