SO 5631-4784: Using setitimer()
The code for working with `setitimer()` and `gettimer()` now seems to be
working reasonably sanely, but the ructions in my main library to
support it were considerable.  I needed arithmetic for `struct timeval`
(since the `struct itimerval` embeds two `struct timeval` values).  This
is why these are obsolescent in the POSIX specification — the move is
to make all `struct timeval` functions obsolete in favour of `struct
timespec`.  But there then turned out to be compilation issues with test
programs for the code in the library after upgrading to GCC 9.1.0
(primarily issues with `-Wstringop-truncation` for `strncat()` and
`strncpy()`, and issues `-Wcast-qual` and string literals initializing
non-const `char` pointers), so those needed to be fixed, and some of the
fixes are non-trivial. Grump!