SO 4295-4541 — Sorting odd or even valued elements of an array
The perceived problem is to take an array of integers, both odd and even values,
and to sort either the odd or the even values into ascending order, while not
moving the values with the other parity in the array.

The code to do so using an auxilliary array as big as the data array (but not
copying the data into that array) is developed incrementally.

The code to do so without the auxilliary array is also shown, but it is subtle
and slow and not recommended.