Make changes as suggested by Peter Cordes
Change the "Not so good way to declare variables" section to use long
instead of int, as suggested by Peter Cordes
https://stackoverflow.com/users/224132/peter-cordes.

However, introducing the alternative variable name, shared_var, that
occurs nowhere in the code is confusing (to me), so change the code to
use 'long l' instead of 'int i' — and it is a damn good job that I
check compile the code.

Add the note about GCC 10.1.0 no longer linking the 'common' example
by default.  It works OK with GCC 9.3.0 but not 10.1.0.