Issue #222 - less zealous flattening
The `flatten.pl` script flattened a paragraph of the `prog1.mk` makefile
because it contained a comment with 6 consecutive words.  Revised to
ignore paragraphs that contain a line starting with a hash `#`.  That is
probably adequate for this question.  As a general purpose tool, it is
probably not sufficient; be cautious about using the code elsewhere.