src/matrices/matmul99.c - revised
The old matrix multiplication function is still present, renamed to
matrix_multiply_arcane().  A replacement version which simulates how
you'd multiply two matrices by hand is added as matrix_multiply().
The code uses double rather than float throughout - including in the
matrix initializers.

The difference in the calculation is that the roles of the two inner
loops are reversed, and by using the sum variable, avoids writing to
any cell in the output matrix more than once.  One consequence of this
is that the result matrix does not need to be zeroed first.