SO 0612-7503: Add check-conservation.[ch]
The code checks whether a recently added shuffle algorithm conserves the
input data when shuffling the array.  The function
`check_conservation()` does the necessary checking, and the
infrastructure validates the results.  In the process, it demonstrates
that the new algorithm does NOT conserve the input data, in two
different ways (it creates the output in a second array, but destroys
the input array during the processing; the output array does not
conserve the data in the original input array).  There's a chance that
the answer will be deleted, requiring 10k reputation to see it.