SO 1974-7644: Support for SO 7430-0464 too.
The numproc19.c program doesn't use any JLSS library facilities, unlike
numproc23.c.  It works adequately, though there's room for nitpicking
with the random number generation (it's not great; it is adequate for
demo purposes).  There are notes on how the code might be productized.

The program numproc23.c is still here.  It won't compile unless you have
prng48.[ch] and randseed.[ch], so it is not compiled by default.  It
also uses functions provided in the src/libsoq directory.  It is not
hard to see how numproc19.c was carved out of numproc23.c.