SO 0612-7503: Update check-conservation.c; add compare-arrays.c
The code in check-conservation.c runs sanely to the ends of the
arrays, but is for data of type int.

The code in compare-arrays.c is for a generic data type identified
by void pointer base address, number of elements and size of each
element.  It is missing the 'once one array is used up' code that
is present in check-conservation.c