SO 0612-0753: Update check-conservation.c and compare-arrays.c
Minor tweaks to check-conservation.c: use a while loop instead
of a partial for loop.  Revise tail loops to avoid two increments.

Major work on compare-arrays.c: fix the main comparator loop in
print_array_differences() and add the two tail loops.  Upgrade
test_conservation() to call print_array_differences() if there
is a difference.  However, the genericity is incomplete (to be
polite about it).  Fixing that is next.