SO 0612-7503: Update compare-arrays.c
The code for test_conservation() is now fully genericized.  There is a
good basis for adding test_conservation(), print_sorted_copy_of_array(),
print_array_differences() to the check-conservation.h header.  Well,
almost.  The 'print_sorted_copy_of_array()' should probably be split
into separate functions: dump_generic_array() (already exists) and ...
well, copy generic array() is memmove(), sort generic array is qsort(),
and the printing in the function is of the original array and the sorted
copy.  And it calls check_conservation() for good measure, but there's
really no cause for it to do so as qsort() won't go wrong given accurate
parameters.  The code in print_array_differences should probably use
array_entry() or array_entry_c().  There are various test artefacts such
as cmp_int() and format_int() which could be renamed (compare_int()) and
published.  An argument to print_array_differences() could configure it
not to print equality.

However, in general, the code would be used in testing where verbosity
is not abhorrent - it helps show where there are bugs in shuffle or sort
code.

The code in compare-arrays.c needs to be validated with a data type
other than integer, but that is mostly a formality, I think.  That
requires at least one extra compare and format pair of functions.  It's
possible that the formatters should be given an extra "void *ctxt"
argument so that things like format strings can be passed to the
formatting function.