SO 3442-6734: Add the question to the README file
The question was deleted on 2016-02-17.  The question did not show any
effort, and the answers were equally evasive.  However, I coded up a
decent working solution.  It features a Fisher-Yates shuffle, and tests
cycles of lengths that can be controlled by command line arguments.

This set of changes adds the question and my answer to the README file.
I added what would go in the help message if the `-h` option were not
being used for the high value in the range of lengths.

Usage: set-sk [-h hi][-l lo][-r repeat][-s seed][-v]

   -h hi      Longest array length (default 15)
   -l lo      Shortest array length (default 5)
   -r repeat  Number of times to shuffle the array (default 1)
   -s seed    Use the given number as the random seed
   -v         Verbose mode

The other not-very-helpful answer suggests a much more complex solution,
and I'm not even sure it would work.  There was a comment by the person
answering that "first, this is NP hard problem".  That is nonsense; it
can be solved in O(N^2) time, which is not NP.