src/miscellany: add f2c.bc
Script for use with bc to provide functions f2c to convert from a
temperature in degrees Fahrenheit to degrees Centigrade or Celsius, and
c2f to convert from Celsius to Fahrenheit.

The conversions use the symmetric formula (x + 40) * scale - 40, where
scale is 5/9 for Fahrenheit to Celsius and 9/5 for Celsius to
Fahrenheit.