src/libsoq: add stderrloc.[ch] and noreturn.h
The header stderrloc.h provides convenient interfaces to funtions which
report on source file, line number and file name as well as the rest of
the error message.  The code uses the error reporting functions from
stderr.[ch] to ensure that conventions such as time and PID stamping
are followed too.