src/so-0202-7991: Add answer to SO 0202-7991
Lists of headers for C and C++ under various standards.

The supporting data needs working up better for C++.
There are the C++ headers such as <iostream>.
There are the C headers such as <cstdio>
There are the .h headers such as <stdio.h>
They should be in c++XY-headers, c++XY-c-headers and c++XY-h-headers.
The -h-headers files are all missing.
The -c-headers files are spotty.
The answer text needs to be be made consistent.
The tables in the answer should be formatted with:

    multicol -c 4 -f source