src/so-4005-2393
Not a good question, and I didn't contribute an answer.  The program
ll31 crashes in function remove_node1() for reasons which are not clear.
Maybe, some day, I'll get to the bottom of this.

Cleaned up calls to calloc() to have 'number' then 'size' rather than
the other way around.