SO 0164-4868: #define macro for debug printing in C?
It's embarrassing that this answer was not in my SOQ repo for so long.
The question was asked on 2009-10-29; my answer was posted the same day
(remember FGITW)?  It's been through multiple revisions; I've not
attempted to recreate that history.

There is, as yet, no source code here except embedded in the answer.